+++ web-wallet/src/lib/dusk/components/__tests__/Mnemonic.spec.js
@@ -44,7 +44,7 @@ describe("Mnemonic", () => {
 	});
 
 	it("should display all the words in the order they have been clicked", async () => {
-		const { getAllByRole } = render(Mnemonic, {
+		const { container, getAllByRole } = render(Mnemonic, {
 			props: {
 				enteredMnemonicPhrase: enteredSeed,
 				mnemonicPhrase: seed,
@@ -52,7 +52,7 @@ describe("Mnemonic", () => {
 			}
 		});
 
-		const buttons = getAllByRole("button");
+		const buttons = container.querySelectorAll("button[data-value]");
 
 		for (const word of buttons) {
 			await fireEvent.click(word);
@@ -64,5 +64,41 @@ describe("Mnemonic", () => {
 		enteredPhrase.forEach((word, index) => {
 			expect(word.textContent).toBe(buttons[index].textContent);
 		});
+	});
+
+	it("should revert the most recent word on Undo click", async () => {
+		const { container, getByText, getAllByRole } = render(Mnemonic, {
+			props: {
+				enteredMnemonicPhrase: enteredSeed,
+				mnemonicPhrase: seed,
+				type: "validate"
+			}
+		});
+
+		const buttons = container.querySelectorAll("button[data-value]");
+
+		// Enters the first 5 words
+		for (let i = 0; i < 5; i++) {
+			await fireEvent.click(buttons[i]);
+			expect(buttons[i]).toBeDisabled();
+		}
+
+		const undoButton = getByText("Undo");
+
+		// Presses the "Undo button"
+		await fireEvent.click(undoButton);
+
+		const enteredPhrase = getAllByRole("listitem");
+
+		// Verify that the content of each list item matches the corresponding button's text content
+		// Loop until the 4th item (index 3)
+		for (let i = 0; i <= 3; i++) {
+			expect(enteredPhrase[i].textContent).toBe(buttons[i].textContent);
+		}
+
+		// The 5th item (index 4) and next should be empty
+		for (let i = 4; i <= 11; i++) {
+			expect(enteredPhrase[i].textContent).toBe("_____");
+		}
 	});
 });